codespace: Renamed Renderer/ -> Graphics/, Object/ -> Graphics/; bug: Screen still empty

+++ VolcaniCore/src/impl/OpenGL/Renderer.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <Renderer/RendererAPI.h>
+#include <Graphics/RendererAPI.h>
 
 namespace VolcaniCore::OpenGL {
 
@@ -15,7 +15,7 @@ public:
 	DrawBuffer* NewDrawBuffer(DrawBufferSpecification& specs,
 							  void* data = nullptr) override;
 	DrawBuffer* GetDrawBuffer(DrawBufferSpecification& specs) override;
-	void SetBufferData(DrawBuffer* buffer, uint32_t bufferIndex,
+	void SetBufferData(DrawBuffer* buffer, uint8_t bufferIndex,
 		const void* data, uint64_t count, uint64_t offset = 0) override;
 	void ReleaseBuffer(DrawBuffer* buffer) override;
